JavadocReplacementLogicおよびJavadocReplacementLogicImplクラスを削除し、JdaReplacementModelImplクラスにJavadoc処理ロジックを統合。コードの可読性とメンテナンス性を向上。
